+++ task_generator/task_generator/constants.py
@@ -1,4 +1,6 @@
+import dataclasses
 from enum import Enum
+import enum
 import math
 import random
 from typing import Any, Callable, List, Optional, TypeVar, TypedDict
@@ -24,67 +26,115 @@ class Constants:
         PEDSIM = "pedsim"
         FLATLAND = "flatland"
 
-    class TaskMode(Enum):
-        RANDOM = "random"
-        STAGED = "staged"
-        SCENARIO = "scenario"
-        PARAMETRIZED = "parametrized"
-        DYNAMIC_MAP_RANDOM = "dynamic_map_random"
-        DYNAMIC_MAP_STAGED = "dynamic_map_staged"
-        GUIDED = "guided"
-        EXPLORE = "explore"
+    class TaskMode:
+
+        @enum.unique
+        class TM_Obstacles(enum.Enum):
+            PARAMETRIZED = "parametrized"
+            RANDOM = "random"
+            SCENARIO = "scenario"
+
+        @enum.unique
+        class TM_Robots(enum.Enum):
+            GUIDED = "guided"
+            EXPLORE = "explore"
+            RANDOM = "random"
+            SCENARIO = "scenario"
+
+        @enum.unique
+        class TM_Module(enum.Enum):
+            STAGED = "staged"
+            DYNAMIC_MAP = "dynamic_map"
+            CLEAR_FORBIDDEN_ZONES = "clear_forbidden_zones"
+            RVIZ_UI = "rviz_ui"
+
+    class MapGenerator:
+        NODE_NAME = "map_generator"
+        MAP_FOLDER_NAME = "dynamic_map"
+
+    PLUGIN_FULL_RANGE_LASER = {
+        "type": "Laser",
+        "name": "full_static_laser",
+        "frame": "full_laser",
+        "topic": "full_scan",
+        "body": "link_base",
+        "broadcast_tf": "true",
+        "origin": [0, 0, 0],
+        "range": 2.0,
+        "angle": {"min": -3.14, "max": 3.14, "increment": 0.01745},
+        "noise_std_dev": 0.0,
+        "update_rate": 10,
+    }
 
 
-def reconfigure():
-    class Config:
-        class General:
-            WAIT_FOR_SERVICE_TIMEOUT = rosparam_get(float, "timeout_wait_for_service", 60)  # 60 secs
-            
-        class Robot:
-            GOAL_TOLERANCE_RADIUS = rosparam_get(float, "goal_radius", 1.0)
-            GOAL_TOLERANCE_ANGLE = rosparam_get(float, "goal_tolerance_angle", 30 * math.pi / 180)
-            TIMEOUT = rosparam_get(float, "timeout", 3*60)  # 3 min
-            MAX_RESET_FAIL_TIMES = rosparam_get(int, "max_reset_fail_times", 10)
-            SPAWN_ROBOT_SAFE_DIST = 0.25
+### TaskConfig
 
-        class Obstacles:
-            MIN_DYNAMIC_OBS = rosparam_get(int, "~configuration/task_mode/random/interactive/min", 0)
-            MAX_DYNAMIC_OBS = rosparam_get(int, "~configuration/task_mode/random/interactive/max", 0)
+@dataclasses.dataclass
+class TaskConfig_General:
+    WAIT_FOR_SERVICE_TIMEOUT: float
+    MAX_RESET_FAIL_TIMES: int
 
-            MIN_STATIC_OBS = rosparam_get(int, "~configuration/task_mode/random/static/min", 0)
-            MAX_STATIC_OBS = rosparam_get(int, "~configuration/task_mode/random/static/max", 0)
+@dataclasses.dataclass
+class TaskConfig_Robot:
+    GOAL_TOLERANCE_RADIUS: float
+    GOAL_TOLERANCE_ANGLE: float
+    SPAWN_ROBOT_SAFE_DIST: float
+    TIMEOUT: float
 
-            MIN_INTERACTIVE_OBS = rosparam_get(int, "~configuration/task_mode/random/interactive/min", 0)
-            MAX_INTERACTIVE_OBS = rosparam_get(int, "~configuration/task_mode/random/interactive/max", 0)
+@dataclasses.dataclass
+class TaskConfig_Obstacles:
+    MIN_DYNAMIC_OBS: int
+    MAX_DYNAMIC_OBS: int
 
-            MODELS_DYNAMIC_OBSTACLES: List[str] = rosparam_get(list, "~configuration/task_mode/random/dynamic/models", [])
-            MODELS_INTERACTIVE_OBSTACLES: List[str] = rosparam_get(list, "~configuration/task_mode/random/interactive/models", [])
-            MODELS_STATIC_OBSTACLES: List[str] = rosparam_get(list, "~configuration/task_mode/random/static/models", [])
+    MIN_STATIC_OBS: int
+    MAX_STATIC_OBS: int
+
+    MIN_INTERACTIVE_OBS: int
+    MAX_INTERACTIVE_OBS: int
+
+    MODELS_DYNAMIC_OBSTACLES: List[str]
+    MODELS_INTERACTIVE_OBSTACLES: List[str]
+    MODELS_STATIC_OBSTACLES: List[str]
+
+    OBSTACLE_MAX_RADIUS: float
+
+@dataclasses.dataclass
+class TaskConfig:
+    General: TaskConfig_General
+    Robot: TaskConfig_Robot
+    Obstacles: TaskConfig_Obstacles
+
+
+def reconfigure() -> TaskConfig:
+
+    return TaskConfig(
+        General=TaskConfig_General(
+            WAIT_FOR_SERVICE_TIMEOUT = rosparam_get(float, "timeout_wait_for_service", 60),  # 60 secs
+            MAX_RESET_FAIL_TIMES = rosparam_get(int, "max_reset_fail_times", 10),
+        ),
+        Robot=TaskConfig_Robot(
+            GOAL_TOLERANCE_RADIUS = rosparam_get(float, "goal_radius", 1.0),
+            GOAL_TOLERANCE_ANGLE = rosparam_get(float, "goal_tolerance_angle", 30 * math.pi / 180),
+            SPAWN_ROBOT_SAFE_DIST = 0.25,
+            TIMEOUT = rosparam_get(float, "timeout", 3*60),  # 3 min
+        ),
+        Obstacles=TaskConfig_Obstacles(
+            MIN_DYNAMIC_OBS = rosparam_get(int, "~configuration/task_mode/random/interactive/min", 0),
+            MAX_DYNAMIC_OBS = rosparam_get(int, "~configuration/task_mode/random/interactive/max", 0),
+
+            MIN_STATIC_OBS = rosparam_get(int, "~configuration/task_mode/random/static/min", 0),
+            MAX_STATIC_OBS = rosparam_get(int, "~configuration/task_mode/random/static/max", 0),
+
+            MIN_INTERACTIVE_OBS = rosparam_get(int, "~configuration/task_mode/random/interactive/min", 0),
+            MAX_INTERACTIVE_OBS = rosparam_get(int, "~configuration/task_mode/random/interactive/max", 0),
+
+            MODELS_DYNAMIC_OBSTACLES = rosparam_get(list, "~configuration/task_mode/random/dynamic/models", []),
+            MODELS_INTERACTIVE_OBSTACLES = rosparam_get(list, "~configuration/task_mode/random/interactive/models", []),
+            MODELS_STATIC_OBSTACLES = rosparam_get(list, "~configuration/task_mode/random/static/models", []),
 
             OBSTACLE_MAX_RADIUS = 15
-
-        class MapGenerator:
-            NODE_NAME = "map_generator"
-            MAP_FOLDER_NAME = "dynamic_map"
-            
-        class Scenario:
-            RESETS_DEFAULT = 5
-
-        PLUGIN_FULL_RANGE_LASER = {
-            "type": "Laser",
-            "name": "full_static_laser",
-            "frame": "full_laser",
-            "topic": "full_scan",
-            "body": "link_base",
-            "broadcast_tf": "true",
-            "origin": [0, 0, 0],
-            "range": 2.0,
-            "angle": {"min": -3.14, "max": 3.14, "increment": 0.01745},
-            "noise_std_dev": 0.0,
-            "update_rate": 10,
-        }
-
-    return Config
+        )
+    )
 
 Config = reconfigure()
